split image and raster
