+++ rubato/struct/gameobject/sprites/__init__.py
@@ -1,4 +1,5 @@
 """Holds all image-related components."""
-from .surf_comps import Image, Raster
+from .raster import Raster
+from .image import Image
 from .animation import Animation
 from .spritesheet import Spritesheet
